Move getFeaturedImageMessage to helpers
It’s not actually a UI component, just a string, which is better suited here.
