+++ assets/src/components/index.js
@@ -8,7 +8,6 @@ export { default as StoryBlockDropZone } from './story-block-drop-zone';
 export { default as StoryBlockMover } from './block-mover';
 export { default as FontFamilyPicker } from './font-family-picker';
 export { default as hasMinimumStoryPosterDimensions } from './has-minimum-story-poster-dimensions';
-export { default as getFeaturedImageMessage } from './get-featured-image-message';
 export { default as getFeaturedImageNotice } from './get-featured-image-notice';
 export { default as getPrePublishNotice } from './get-pre-publish-notice';
 export { default as withAmpStorySettings } from './with-amp-story-settings';
