+++ assets/src/helpers.js
@@ -10,7 +10,8 @@ import uuid from 'uuid/v4';
  */
 import { render } from '@wordpress/element';
 import { count } from '@wordpress/wordcount';
-import { _x } from '@wordpress/i18n';
+import { __, _x } from '@wordpress/i18n';
+import { select } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -430,4 +431,26 @@ export const getPercentageFromPixels = ( axis, pixelValue ) => {
 		return Math.round( ( pixelValue / STORY_PAGE_INNER_HEIGHT ) * 100 );
 	}
 	return 0;
+};
+
+/**
+ * Gets the message for an invalid featured image, if it is invalid.
+ *
+ * @param {Function} validateImageSize A function to validate whether the size is correct.
+ * @param {string} invalidSizeMessage A message to display in a Notice if the size is wrong.
+ * @return {string|null} A message about the invalid featured image, or null if it's valid.
+ */
+export const getFeaturedImageMessage = ( validateImageSize, invalidSizeMessage ) => {
+	const currentPost = select( 'core/editor' ).getCurrentPost();
+	const editedFeaturedMedia = select( 'core/editor' ).getEditedPostAttribute( 'featured_media' );
+	const featuredMedia = currentPost.featured_media || editedFeaturedMedia;
+
+	if ( ! featuredMedia ) {
+		return __( 'Selecting a featured image is required.', 'amp' );
+	}
+
+	const media = select( 'core' ).getMedia( featuredMedia );
+	if ( ! media || ! media.media_details || ! validateImageSize( media.media_details ) ) {
+		return invalidSizeMessage;
+	}
 };
